refactor(dashboard): consolidate admin and workers components into a single reusable component
This commit introduces a new reusable `WorkersAndAdmin` component to handle both admin and worker functionalities, reducing code duplication. The existing `ModalAdminDashboard` and `ModalWorkersDashboard` components have been merged into `ModalWorkersAndAdminDashboard`. Additionally, the `user.slice.ts` has been updated to include actions for managing admins. This refactor improves maintainability and reduces redundancy in the codebase.
