+++ src/app/dashboard/workers/page.tsx
@@ -1,138 +1,10 @@
 "use client";
-import { useAppDispatch, useAppSelector } from "@/app/hooks";
-import DashboardLayout from "@/components/dashboard/dashboardLayout";
-import ModalWorkersDashboard from "@/components/dashboard/workers/modalWorkers.dashboard";
-import ModalLayoutComponent from "@/components/layouts/modal.layout";
-import ErrorSharedComponent from "@/components/shared/error.shared.component";
-import { initialWorkers, removeWorkers } from "@/feature/user.slice";
+import WorkersAndAdmin from "@/components/dashboard/workersAndAdmin/tableWorkerAdmin.dashboard";
 import { withAuth } from "@/hoc/withAuth";
-import useConfirmDialog from "@/hooks/useConfirmDialog.hook";
-import { useDataFetch } from "@/hooks/useDataFetch.hook";
-import { User } from "@/type";
-import { fetchData } from "@/utils/fetchData.util";
-import { FC, useEffect, useRef, useState } from "react";
+import { FC } from "react";
 
 const Workers: FC = () => {
-  const [data, loading] = useDataFetch<User[]>(
-    "/user/worker",
-    false,
-    0,
-    0,
-    true
-  );
-  const workers = useAppSelector((state) => state.user.workers);
-  const dispatchApp = useAppDispatch();
-  const [isModalOpen, setIsModalOpen] = useState<boolean>(false);
-  const modalRef = useRef<HTMLDivElement>(null);
-  const [error, setError] = useState<string | string[]>("");
-  const { confirm, ConfirmDialog } = useConfirmDialog();
-
-  useEffect(() => {
-    if (!loading) {
-      dispatchApp(initialWorkers(data));
-    }
-    return () => {};
-  }, [data, dispatchApp, loading]);
-
-  const handlerDelete = async (id: string) => {
-    const accepted = await confirm({
-      title: "Delete user?",
-      description:
-        "This can't be undone. You are about to permanently delete something",
-      confirmText: "Yes, delete",
-      cancelText: "Cancel",
-    });
-
-    if (!accepted) return;
-
-    try {
-      const data = await fetchData<User>(
-        `/user/${id}`,
-        {},
-        "DELETE",
-        localStorage.getItem("token") || ""
-      );
-
-      if (data.message === undefined) {
-        dispatchApp(removeWorkers(data));
-      }
-    } catch (error) {
-      if (error instanceof Error) {
-        setError(error.message);
-      }
-    }
-  };
-
-  return (
-    <DashboardLayout selected="Workers">
-      {ConfirmDialog}
-      <ErrorSharedComponent error={error} />
-      <>
-        {isModalOpen && (
-          <ModalLayoutComponent modalRef={modalRef}>
-            <ModalWorkersDashboard setIsModalOpen={setIsModalOpen} />
-          </ModalLayoutComponent>
-        )}
-        <div className="p-6">
-          <div className="flex justify-between items-center mb-6">
-            <h1 className="text-2xl font-semibold text-secondary-800">
-              Workers
-            </h1>
-            <button
-              onClick={() => setIsModalOpen(true)}
-              className="bg-primary-500 hover:bg-primary-600 text-white px-4 py-2 rounded-lg transition-colors cursor-pointer"
-            >
-              Add New Worker
-            </button>
-          </div>
-
-          <div className="bg-white rounded-lg shadow-md overflow-hidden">
-            <table className="w-full">
-              <thead className="bg-secondary-50">
-                <tr>
-                  <th className="px-6 py-3 text-left text-sm font-medium text-secondary-700">
-                    Name
-                  </th>
-                  <th className="px-6 py-3 text-left text-sm font-medium text-secondary-700">
-                    Email
-                  </th>
-                  <th className="px-6 py-3 text-left text-sm font-medium text-secondary-700">
-                    Role
-                  </th>
-                  <th className="px-6 py-3 text-left text-sm font-medium text-secondary-700">
-                    Actions
-                  </th>
-                </tr>
-              </thead>
-              <tbody className="divide-y divide-secondary-100">
-                {workers.map((worker) => (
-                  <tr key={worker.email} className="hover:bg-secondary-50">
-                    <td className="px-6 py-4 text-sm text-secondary-800">
-                      {worker.name}
-                    </td>
-                    <td className="px-6 py-4 text-sm text-secondary-800">
-                      {worker.email}
-                    </td>
-                    <td className="px-6 py-4 text-sm text-secondary-800">
-                      {worker.role}
-                    </td>
-                    <td className="px-6 py-4 text-sm">
-                      <button
-                        onClick={() => handlerDelete(worker.id)}
-                        className="text-accent-error hover:text-red-700 cursor-pointer"
-                      >
-                        Delete
-                      </button>
-                    </td>
-                  </tr>
-                ))}
-              </tbody>
-            </table>
-          </div>
-        </div>
-      </>
-    </DashboardLayout>
-  );
+  return <WorkersAndAdmin type="worker" />;
 };
 
 export default withAuth(Workers);
